# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant Management System API",
        "description": "API documentation for Restaurant Management System Backend",
        "contact": {
            "email": "admin@restaurant.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Restaurant Management API Server"
        }
    ],
    "paths": {
...
        "/api/auth/dining-tables": {
            "get": {
                "tags": [
                    "DiningTables"
                ],
                "summary": "Get all dining tables",
                "description": "Retrieve all dining tables with pagination",
                "operationId": "c7a2c719935f10e5301beb68112f5324",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dining tables retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dining tables retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "example": "DT001"
                                                            },
                                                            "table_number": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "capacity": {
                                                                "type": "integer",
                                                                "example": 4
                                                            },
                                                            "is_active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DiningTables"
                ],
                "summary": "Create new dining table",
                "description": "Create a new dining table",
                "operationId": "ff12136906499febd175df822b491a17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_number",
                                    "capacity"
                                ],
                                "properties": {
                                    "table_number": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dining table created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/dining-tables/{id}": {
            "get": {
                "tags": [
                    "DiningTables"
                ],
                "summary": "Get dining table by ID",
                "description": "Retrieve a specific dining table by ID",
                "operationId": "efc819ea6a824888cf98851a0b647ad3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dining Table ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dining table retrieved successfully"
                    },
                    "404": {
                        "description": "Dining table not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "DiningTables"
                ],
                "summary": "Update dining table",
                "description": "Update an existing dining table",
                "operationId": "ec139060fa0f40ad6ccb7afa133a91dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dining Table ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "table_number": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dining table updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DiningTables"
                ],
                "summary": "Delete dining table",
                "description": "Delete a dining table",
                "operationId": "c634b6b7741425ce9608a70fd5b1d644",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dining Table ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dining table deleted successfully"
                    },
                    "404": {
                        "description": "Dining table not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}